+++ lib/flutter_admob_app_open.dart
@@ -1,8 +1,11 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:firebase_admob/firebase_admob.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+
+export 'package:firebase_admob/firebase_admob.dart';
 
 class FlutterAdmobAppOpen {
   static final FlutterAdmobAppOpen _instance = FlutterAdmobAppOpen._();
@@ -28,12 +31,14 @@ class FlutterAdmobAppOpen {
   Future<bool> initialize({
     @required String appId,
     String appAppOpenAdUnitId,
+    MobileAdTargetingInfo targetingInfo,
   }){
     assert(appId != null && appId.isNotEmpty);
 
     return _channel.invokeMethod<bool>("initialize", <String, dynamic>{
       'appId': appId,
       'appAppOpenAdUnitId': appAppOpenAdUnitId,
+      'targetingInfo': targetingInfo?.toJson(),
     });
   }
 }