+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.5
+
+* Support MobileAdTargetingInfo from firebase_admob. You can use MobileAdTargetingInfo to target adslike keywords, childDirected, testDevices, ...
+* Fixed duplicate ads on android.
+
 ## 0.0.4
 
 * Support iOS.